+++ src/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
 import requests from './requestsReducer';
+import response from './responseReducer';
 
 export default combineReducers({
-  requests: requests
+  requests,
+  response
 });
